--- roomrando/roomrando.py
@@ -1,364 +1,0 @@
-# External libraries
-import os
-import yaml
-
-class Address:
-    '''
-    # gamedata address:
-    #   - The location of a specific piece of gamedata
-    #   - Used as the canonical address
-    # disc address:
-    #   - The location of gamedata as it is found on the disc image
-    #   - To get the location on disc, call get_disc_address passing the 
-    #     canonical or gamedata address as a parameter
-    '''
-    SECTOR_HEADER_SIZE = 24
-    SECTOR_DATA_SIZE = 2048
-    SECTOR_ERROR_CORRECTION_DATA_SIZE = 280
-    SECTOR_SIZE = SECTOR_HEADER_SIZE + SECTOR_DATA_SIZE + SECTOR_ERROR_CORRECTION_DATA_SIZE
-    def __init__(self, address: int, address_type: str='GAMEDATA'):
-        if address_type == 'DISC':
-            self.address = self.get_gamedata_address(address)
-        elif address_type == 'GAMEDATA':
-            self.address = address
-        else:
-            raise ValueError('Unknown address type: ' + address_type)
-    
-    def to_disc_address(self, offset: int=0) -> int:
-        result = self.get_disc_address(self.address + offset)
-        return result
-
-    @classmethod
-    def get_disc_address(self, gamedata_address: int) -> int:
-        sector, offset = divmod(gamedata_address, self.SECTOR_DATA_SIZE)
-        result = sector * self.SECTOR_SIZE + self.SECTOR_HEADER_SIZE + offset
-        return result
-
-    @classmethod
-    def get_gamedata_address(self, disc_address: int) -> int:
-        HDR = self.SECTOR_HEADER_SIZE
-        DAT = self.SECTOR_DATA_SIZE
-        sector, offset = divmod(disc_address, self.SECTOR_SIZE)
-        if offset < HDR:
-            return None
-        elif offset >= (HDR + DAT):
-            return None
-        result = sector * DAT + (offset - HDR) % DAT
-        return result
-
-class Room:
-    def __init__(self, room_index: int, box: tuple[int], exits: list[tuple], flags: set[str], foreground_layer_id: int):
-        self.room_index = room_index
-        self.top = box[0]
-        self.left = box[1]
-        self.height = box[2]
-        self.width = box[3]
-        self.exits = exits
-        self.flags = flags
-        self.foreground_layer_id = foreground_layer_id
-
-class Teleporter:
-    def __init__(self, teleporter_index: int, x: int, y: int, room_index: int, current_stage_id: int, next_stage_id: int):
-        self.teleporter_index = teleporter_index
-        self.x = x
-        self.y = y
-        self.room_index = room_index
-        self.current_stage_id = current_stage_id
-        self.next_stage_id = next_stage_id
-
-class IndexedBitmapCanvas():
-    def __init__(self, rows, cols):
-        self.address = Address(0x001AF800, 'GAMEDATA')
-        self.rows = rows
-        self.cols = cols
-        self.pixels = [
-            [None for col in range(self.cols)] for row in range(self.rows)
-        ]
-        self.default = 1
-    
-    def get_pixel(self, top: int, left: int):
-        result = self.pixels[top][left]
-        return result
-    
-    def set_pixel(self, top: int, left: int, fill: int):
-        # TODO(sestren): Fix the occasional IndexError: list index out of range when calling draw_room()
-        self.pixels[top][left] = fill
-    
-    def fill_rect(self, top: int, left: int, height: int, width: int, fill: int):
-        for row in range(height):
-            for col in range(width):
-                self.set_pixel(top + row, left + col, fill)
-    
-    def draw_room(self, room: Room):
-        top = 4 * room.top
-        left = 4 * room.left
-        height = 4 * room.height + 1
-        width = 4 * room.width + 1
-        self.fill_rect(top, left, height, width, 0)
-        self.fill_rect(top + 1, left + 1, height - 2, width - 2, 1)
-        for (row, col, edge) in room.exits:
-            y = top + 4 * row + 2
-            x = left + 4 * col + 2
-            if edge == 'Left':
-                x -= 2
-            elif edge == 'Right':
-                x += 2
-            elif edge == 'Top':
-                y -= 2
-            elif edge == 'Bottom':
-                y += 2
-            self.set_pixel(y, x, 1)
-
-class PPF:
-    def __init__(self, description):
-        self.patches = []
-        self.description = (description + 50 * ' ')[:50]
-        self.bytes = bytearray()
-        self.write_string('PPF30')
-        self.write_byte(2) # Encoding method = PPF3.0
-        self.write_string(self.description)
-        self.write_byte(0) # Imagetype = BIN
-        self.write_byte(0) # Blockcheck = Disabled
-        self.write_byte(0) # Undo data = Not available
-        self.write_byte(0) # Dummy
-        assert len(self.bytes) == 60 # 0x3C
-    
-    def write_byte(self, byte):
-        assert 0x00 <= byte < 0x100
-        self.bytes.append(byte)
-    
-    def write_string(self, string):
-        for char in string:
-            self.write_byte(ord(char))
-    
-    def write_u16(self, value):
-        for _ in range(2):
-            value, byte = divmod(value, 0x100)
-            self.write_byte(byte)
-    
-    def write_u32(self, value):
-        for _ in range(4):
-            value, byte = divmod(value, 0x100)
-            self.write_byte(byte)
-    
-    def write_u64(self, value):
-        for _ in range(8):
-            value, byte = divmod(value, 0x100)
-            self.write_byte(byte)
-    
-    def patch_string(self, offset_in_file: int, value: str):
-        self.write_u64(offset_in_file)
-        size = len(value)
-        self.write_byte(size)
-        self.write_string(value)
-    
-    def patch_bitmap(self, canvas: IndexedBitmapCanvas, address: Address):
-        for row in range(0, canvas.rows):
-            # Each byte contains a pair of pixels, side-by-side horizontally
-            for col in range(0, canvas.cols, 2):
-                left = canvas.get_pixel(row, col)
-                right = canvas.get_pixel(row, col + 1)
-                # No overwrites on either pair of pixels, so skip them
-                if left is None and right is None:
-                    continue
-                disc_address = address.to_disc_address(
-                    row * (canvas.cols // 2) + col // 2
-                )
-                self.write_u64(disc_address)
-                size = 1
-                self.write_byte(size)
-                # If either side of the pair is unspecified, supply the default
-                left = 0xF & (left if left is not None else canvas.default)
-                right = 0xF & (right if right is not None else canvas.default)
-                self.write_byte((right << 4) | left)
-    
-    def patch_room_data(self, room: Room, address: Address):
-        self.write_u64(address.to_disc_address(8 * room.room_index))
-        size = 4
-        self.write_byte(size)
-        self.write_byte(room.left)
-        self.write_byte(room.top)
-        self.write_byte(room.left + room.width - 1)
-        self.write_byte(room.top + room.height - 1)
-    
-    def patch_teleporter_data(self, teleporter: Teleporter, address: Address):
-        self.write_u64(address.to_disc_address(10 * teleporter.teleporter_index))
-        size = 10
-        self.write_byte(size) 
-        self.write_u16(teleporter.x)
-        self.write_u16(teleporter.y)
-        self.write_u16(8 * teleporter.room_index)
-        self.write_u16(teleporter.current_stage_id)
-        self.write_u16(teleporter.next_stage_id)
-    
-    def patch_packed_room_data(self, room: Room, address: Address):
-        write_address = address.to_disc_address(0x10 * room.foreground_layer_id + 0x08)
-        self.write_u64(write_address)
-        size = 4
-        self.write_byte(size)
-        flags_byte = 0x00
-        flags = {
-            'Load On Left': 0x40,
-            'Load On Right': 0x41,
-            'Save With Left Opening': 0x20,
-            'Save With Right Opening': 0x22,
-            'Scroll Type 1': 0x01,
-            'Special Type 1': 0x80,
-            'Special Type 2': 0x92,
-        }
-        for flag_name in room.flags:
-            if flag_name in flags:
-                flags_byte |= flags[flag_name]
-        data = [
-            flags_byte,
-            0x3F & (room.top + room.height - 1), # bottom
-            0x3F & (room.left + room.width - 1), # right
-            0x3F & (room.top),
-            0x3F & (room.left),
-        ]
-        self.write_u32(
-            (data[0] << 24) |
-            (data[1] << 18) |
-            (data[2] << 12) |
-            (data[3] << 6) |
-            (data[4] << 0)
-        )
-
-class DataCore:
-    def __init__(self):
-        print('Build data core')
-        self.rooms = {}
-        self.teleporters = {}
-        for stage_folder in (
-            'castle-entrance',
-            'castle-entrance-revisited',
-            'alchemy-laboratory',
-            'marble-gallery',
-            'outer-wall',
-            'olroxs-quarters',
-        ):
-            folder_path = os.path.join('data', 'rooms', stage_folder)
-            for file_name in os.listdir(folder_path):
-                if file_name[-5:] != '.yaml':
-                    continue
-                file_path = os.path.join(folder_path, file_name)
-                with open(file_path) as open_file:
-                    yaml_obj = yaml.safe_load(open_file)
-                    room_name = yaml_obj['Stage'] + ', ' + yaml_obj['Room']
-                    self.rooms[room_name] = yaml_obj
-        with open(os.path.join('data', 'Teleporters.yaml')) as open_file:
-            yaml_obj = yaml.safe_load(open_file)
-            self.teleporters = yaml_obj
-    
-    def get_core(self) -> dict:
-        result = {
-            'Rooms': self.rooms,
-            'Teleporters': self.teleporters,
-        }
-        return result
-
-class LogicCore:
-    def __init__(self, data_core, changes):
-        print('Build logic core')
-        self.commands = {}
-        for stage_name in (
-            'Castle Entrance',
-            'Castle Entrance Revisited',
-            'Alchemy Laboratory',
-            'Marble Gallery',
-            'Outer Wall',
-            'Olrox\'s Quarters',
-        ):
-            nodes = {}
-            for (location_name, room_data) in data_core['Rooms'].items():
-                if data_core['Rooms'][location_name]['Stage'] != stage_name:
-                    continue
-                room_top = room_data['Top']
-                room_left = room_data['Left']
-                if 'Rooms' in changes and location_name in changes['Rooms']:
-                    if 'Top' in changes['Rooms'][location_name]:
-                        room_top = changes['Rooms'][location_name]['Top']
-                    if 'Left' in changes['Rooms'][location_name]:
-                        room_left = changes['Rooms'][location_name]['Left']
-                self.commands[location_name] = room_data['Commands']
-                for (node_name, node) in room_data['Nodes'].items():
-                    row = room_top + node['Row']
-                    column = room_left + node['Column']
-                    edge = node['Edge']
-                    nodes[(row, column, edge)] = (location_name, node_name, node['Entry Section'])
-                    exit = {
-                        'Outcomes': {
-                            'Location': None,
-                            'Section': None,
-                        },
-                        'Requirements': {
-                            'Default': {
-                                'Location': location_name,
-                                'Section': node['Exit Section']
-                            },
-                        },
-                    }
-                    self.commands[location_name]['Exit - ' + node_name] = exit
-            for (row, column, edge), (location_name, node_name, section_name) in nodes.items():
-                matching_row = row
-                matching_column = column
-                matching_edge = edge
-                if edge == 'Top':
-                    matching_edge = 'Bottom'
-                    matching_row -= 1
-                elif edge == 'Left':
-                    matching_edge = 'Right'
-                    matching_column -= 1
-                elif edge == 'Bottom':
-                    matching_edge = 'Top'
-                    matching_row += 1
-                elif edge == 'Right':
-                    matching_edge = 'Left'
-                    matching_column += 1
-                (matching_location_name, matching_node_name, matching_section) = (None, 'Unknown', None)
-                if (matching_row, matching_column, matching_edge) in nodes:
-                    (matching_location_name, matching_node_name, matching_section) = nodes[(matching_row, matching_column, matching_edge)]
-                self.commands[location_name]['Exit - ' + node_name]['Outcomes']['Location'] = matching_location_name
-                self.commands[location_name]['Exit - ' + node_name]['Outcomes']['Section'] = matching_section
-        # Replace source teleporter locations with their targets
-        for (location_name, location_info) in self.commands.items():
-            for (command_name, command_info) in location_info.items():
-                if 'Outcomes' in command_info and 'Location' in command_info['Outcomes']:
-                    old_location_name = command_info['Outcomes']['Location']
-                    if old_location_name in data_core['Teleporters']['Sources']:
-                        source = data_core['Teleporters']['Sources'][old_location_name]
-                        target = data_core['Teleporters']['Targets'][source['Target']]
-                        new_location_name = target['Stage'] + ', ' + target['Room']
-                        self.commands[location_name][command_name]['Outcomes']['Location'] = new_location_name
-                        target_section_name = data_core['Rooms'][new_location_name]['Nodes'][target['Node']]['Entry Section']
-                        self.commands[location_name][command_name]['Outcomes']['Section'] = target_section_name
-        # Delete fake rooms mentioned as teleporter locations
-        for location_name in data_core['Teleporters']['Sources']:
-            if 'Fake' in location_name:
-                self.commands.pop(location_name, None)
-        self.state = {
-            'Character': 'Alucard',
-            'Location': 'Castle Entrance, After Drawbridge',
-            'Section': 'Ground',
-            'Item - Alucard Sword': 1,
-            'Item - Alucard Shield': 1,
-            'Item - Dragon Helm': 1,
-            'Item - Alucard Mail': 1,
-            'Item - Twilight Cloak': 1,
-            'Item - Necklace of J': 1,
-            'Item - Neutron Bomb': 1,
-            'Item - Heart Refresh': 1,
-        }
-        self.goals = {
-            'Debug - Get Soul of Wolf': {
-                'Relic - Soul of Wolf': True,
-            },
-        }
-    
-    def get_core(self) -> dict:
-        result = {
-            'State': self.state,
-            'Goals': self.goals,
-            'Commands': self.commands,
-        }
-        return result
